FDG-7526 code cleanup
